+++ packages/plugin-node-device/test/device.test.js
@@ -15,7 +15,7 @@ const VALID_NOTIFIER = { name: 't', version: '0', url: 'http://' }
 const ISO_8601 = /^\d{4}(-\d\d(-\d\d(T\d\d:\d\d(:\d\d)?(\.\d+)?(([+-]\d\d:\d\d)|Z)?)?)?)?$/i
 
 describe('plugin: node device', () => {
-  it('should set device = { hostname }, capture the Node version and add a beforeSend callback which adds device time', done => {
+  it('should set device = { hostname, runtimeVersions } add a beforeSend callback which adds device time', done => {
     const client = new Client(VALID_NOTIFIER)
     client.setOptions({ apiKey: 'API_KEY_YEAH' })
     client.configure(schema)